小区名称校验

+++ src/model/data_area.go
@@ -8,7 +8,7 @@ const (
 )
 
 type StationData struct {
-	Uid           int
+	Uid           int64
 	Station_Name  string
 	Province_Name string
 	Province_No   int
@@ -20,23 +20,23 @@ type StationData struct {
 	Station_Phone string
 }
 type OfficeData struct {
-	Uid          int
+	Uid          int64
 	Office_Name  string
-	Station_No   int
+	Station_No   int64
 	Office_Head  string
 	Office_Phone string
 }
 type CommunityData struct {
-	Uid             int
+	Uid             int64
 	Community_Name  string
-	Office_No       int
+	Office_No       int64
 	Community_Head  string
 	Community_Phone string
 }
 type StreetData struct {
 	Uid                   int64
 	Street_Name           string
-	Community_No          int
+	Community_No          int64
 	Street_Head           string
 	Street_Phone          string
 	Street_Property_Name  string
@@ -44,5 +44,5 @@ type StreetData struct {
 }
 type DeleteData struct {
 	Cmd_Delete string
-	Uid        int
+	Uid        int64
 }
